Shift snake only if it has not eaten food
If the snake segments are shifted in the same iteration that food is consumed,
the snake will not grow until the tail passes the location where the
aforementioned food was before being consumed. To allow for the snake to grow
immediately after consuming food, shifting should only occur in an iteration
where no food is eaten.