Flush output buffer after drawing
Flushing the output buffer after drawing fixes a defect where the game ends
immediately upon drawing the head of the snake in a location adjacent to a
point of collision without allowing a change in direction to occur from that
point. Instead, the game should sleep for one iteration, granting the player
a chance to alter the snake's current course to prevent a collision.